// src/api/__generated__/story/model/mintDramaNftRequestPayMethod.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export type MintDramaNftRequestPayMethod =
  (typeof MintDramaNftRequestPayMethod)[keyof typeof MintDramaNftRequestPayMethod];

export const MintDramaNftRequestPayMethod = {
  point: 'point',
  usdt: 'usdt',
  usdc: 'usdc',
} as const;